Remove unnecessary logs, update NewsLetterController and cleaned up ArticleModel
Removed unnecessary logs and switched from using `logException` function to closing the application in the NewsLetterController. A thorough cleanup of the ArticleModel was done, unnecessary `try catch` blocks were removed, and `catid` was added to the queries to improve path generation. Unnecessary spaces and lines were removed throughout the files for cleaner code.