Update SQL scripts with new settings table and trigger
Added a new table `#__semantyca_nm_settings` to house settings data and a trigger `UpdateObsoleteSince` to manage `obsolete_since` field in the install, uninstall, and update SQL scripts. The new table includes a unique constraint for `name` and `obsolete` fields. This trigger sets the `obsolete_since` field when a setting is marked obsolete.