Add statistic retrieval and display enhancements
This commit introduces methods for retrieving statistics about subscriber events related to specific newsletters in the SiteSubscriberEventModel. A column for recipients count is added in the dashboard's statistics tab. A new SubscriberEventController is created to handle request for all subscriber events. Also, recipient data is expanded from JSON format when statistics list is retrieved in StatModel.