Commit message:
Introduce improvements to DynamicFormField.vue and Composer.vue

In DynamicFormField.vue, a debounced search functionality was added for available articles. This provides a better user experience by preventing too many requests from being sent while the user is typing in the search field. Also, a badge was added to show the count of selected articles.

In Composer.vue, the text of the template selection button was changed from "Apply template" to "Select Template" for better clarity.