Commit message:
"Updated styling and functionality of dynamic form field in Joomla project

Updated the styling of draggable items in the dynamic form field for better visual presentation.

Added functionality to clear selected articles with the 'Clear' button and initiate a fetch articles function when a new value is entered in the search field.

These changes enhance the user interaction and usability of the dynamic form field in Joomla project."